# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScriptForge API
  version: 1.0.0
  description: ScriptForge 前台 / 后台 REST API 契约
paths:
...
  /api/admin/model/llm/providers/{provider_id}/activate/:
    post:
      operationId: admin_model_llm_providers_activate_create
      description: POST /api/admin/model/llm/providers/<id>/activate/
      parameters:
      - in: path
        name: provider_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
  /api/admin/model/llm/vendors/{vendor}/credential/:
    put:
      operationId: admin_model_llm_vendors_credential_update
      description: PUT /api/admin/model/llm/vendors/<vendor>/credential/
      parameters:
      - in: path
        name: vendor
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT